Small documentation and naming fixes
Fix some inline documentation, rename request.go, request_test.go to
requests.go, requests_test.go.
