+++ selvpcclient/resell/v2/licenses/schemas.go
@@ -2,12 +2,12 @@ package licenses
 
 import "github.com/selectel/go-selvpcclient/selvpcclient/resell/v2/servers"
 
-// License represents a single Resell Floating IP.
+// License represents a single Resell License.
 type License struct {
 	// ID is a unique id of the license.
 	ID int `json:"id"`
 
-	// ProjectID represents an associated Resell project.
+	// ProjectID represents an associated Identity service project.
 	ProjectID string `json:"project_id"`
 
 	// Region represents a region of where the license resides.
