+++ selvpcclient/resell/v2/roles/requests_opts.go
@@ -7,9 +7,9 @@ type RoleOpts struct {
 
 // RoleOpt represents options for a single Resell role.
 type RoleOpt struct {
-	// ProjectID represents needed Resell project.
+	// ProjectID represents Identity service project.
 	ProjectID string `json:"project_id"`
 
-	// UserID represents needed Resell user.
+	// UserID represents Identity service user.
 	UserID string `json:"user_id"`
 }
