+++ selvpcclient/resell/v2/roles/schemas.go
@@ -1,10 +1,10 @@
 package roles
 
-// Role represents a single Resell subnet.
+// Role represents a single Resell role.
 type Role struct {
-	// ProjectID represents an associated Resell project.
+	// ProjectID represents an associated Identity service project.
 	ProjectID string `json:"project_id"`
 
-	// UserID represents an associated Resell user.
+	// UserID represents an associated Identity service user.
 	UserID string `json:"user_id"`
 }
