+++ selvpcclient/resell/v2/users/requests_opts.go
@@ -8,6 +8,6 @@ type UserOpts struct {
 	// Password represents a user's password.
 	Password string `json:"password,omitempty"`
 
-	// Enabled shows if user is active or it was disabled.
+	// Enabled shows if user is active or it needs to be disabled.
 	Enabled *bool `json:"enabled,omitempty"`
 }
